Implement jwt decoding

+++ src/controllers/recipesController.js
@@ -1,6 +1,8 @@
 const Recipe = require('../models/recipeModel')
 const AppError = require('../utils/appError')
 const catchAsync = require('../utils/catchAsync')
+const jwt = require('jsonwebtoken')
+const { decodeJWT } = require('../utils/checkJwt')
 
 exports.getRecipes = catchAsync(async (req, res, next) => {
     if (
@@ -28,8 +30,12 @@ exports.getRecipes = catchAsync(async (req, res, next) => {
 })
 
 exports.createRecipe = catchAsync(async (req, res, next) => {
+    const token = req.headers.authorization.slice(7)
+    if (!token) return next(new AppError('Not authorized!', 401))
+    const decoded = await jwt.decode(token)
+
     const recipe = await Recipe.create({
-        // userID: req.oidc.sid,
+        userID: decoded.sub,
         ...req.body,
     })
 
@@ -61,11 +67,15 @@ exports.deleteRecipe = catchAsync(async (req, res, next) => {
     const recipeID = req.params.id
     const recipe = await Recipe.findById(recipeID)
 
-    // if (req.oidc.user.sid !== recipe.userID) {
-    //   return next(
-    //     new AppError('Unauthorized to delete a recipe of another user', 401),
-    //   );
-    // }
+    const token = req.headers.authorization.slice(7)
+    if (!token) return next(new AppError('Not authorized!', 401))
+    const decoded = await jwt.decode(token)
+
+    if (decoded.sub !== recipe.userID) {
+        return next(
+            new AppError('Unauthorized to delete a recipe of another user', 401)
+        )
+    }
 
     await Recipe.deleteOne(recipe)
